+++ client/src/pages/HomePage/SubviewsHomePage/FriendsView/FriendsView.css
@@ -55,7 +55,7 @@
     color: #fff;
 }
 
-/* ── SHARED AVATAR ───────────────────────────────────────────────── */
+/* ── SHARED (used by both list and requests tabs) ────────────────── */
 .friend-avatar {
     position: relative;
     width: 36px;
@@ -92,7 +92,6 @@
     color: var(--text);
 }
 
-/* ── FRIENDS SCROLL ──────────────────────────────────────────────── */
 .friends-scroll {
     flex: 1;
     overflow-y: auto;
@@ -104,117 +103,3 @@
     background: rgba(255, 255, 255, 0.1);
     border-radius: 2px;
 }
-
-/* ── FRIENDS GRID ────────────────────────────────────────────────── */
-.friends-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 10px;
-}
-
-.friend-item {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    padding: 8px 10px;
-    border-radius: 10px;
-    background: rgba(255, 255, 255, 0.04);
-    border: 1px solid var(--pb);
-}
-
-.friend-remove {
-    background: transparent;
-    border: none;
-    color: var(--dim);
-    font-family: var(--fb);
-    font-size: 11px;
-    cursor: pointer;
-    transition: color 0.15s;
-    white-space: nowrap;
-}
-
-.friend-remove:hover { color: #f87171; }
-
-/* ── REQUESTS LIST ───────────────────────────────────────────────── */
-.friends-request-list {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-}
-
-.request-item {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    padding: 8px 10px;
-    border-radius: 10px;
-    background: rgba(255, 255, 255, 0.04);
-    border: 1px solid var(--pb);
-}
-
-.request-actions {
-    display: flex;
-    gap: 6px;
-    margin-left: auto;
-}
-
-.req-accept, .req-decline {
-    padding: 4px 12px;
-    border-radius: 6px;
-    border: none;
-    font-family: var(--fb);
-    font-size: 11px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: opacity 0.15s;
-}
-
-.req-accept  { background: var(--cyan);    color: #000; }
-.req-decline { background: rgba(255,255,255,0.08); color: var(--dim); }
-
-.req-accept:hover, .req-decline:hover { opacity: 0.8; }
-
-.friends-empty {
-    color: var(--dim);
-    font-size: var(--font-sm);
-    text-align: center;
-    padding: 20px 0;
-}
-
-/* ── FIND PLAYERS ────────────────────────────────────────────────── */
-.find-row {
-    display: flex;
-    gap: 10px;
-}
-
-.find-input {
-    flex: 1;
-    padding: 8px 14px;
-    border-radius: 8px;
-    border: 1px solid var(--pb);
-    background: rgba(255, 255, 255, 0.05);
-    color: var(--text);
-    font-family: var(--fb);
-    font-size: var(--font-sm);
-    outline: none;
-    transition: border-color 0.15s;
-}
-
-.find-input::placeholder { color: var(--dim); }
-.find-input:focus { border-color: var(--cyan); }
-
-.find-btn {
-    padding: 8px 16px;
-    border-radius: 8px;
-    border: none;
-    background: var(--cyan);
-    color: #000;
-    font-family: var(--fb);
-    font-size: var(--font-sm);
-    font-weight: 700;
-    cursor: pointer;
-    white-space: nowrap;
-    transition: box-shadow 0.2s;
-}
-
-.find-btn:hover { box-shadow: 0 0 16px rgba(0, 229, 255, 0.4); }
