+++ client/src/pages/HomePage/SubviewsHomePage/FriendsView/TabViews/FriendsFindTabView/FriendsFindTabView.css
@@ -1,0 +1,36 @@
+.find-row {
+    display: flex;
+    gap: 10px;
+}
+
+.find-input {
+    flex: 1;
+    padding: 8px 14px;
+    border-radius: 8px;
+    border: 1px solid var(--pb);
+    background: rgba(255, 255, 255, 0.05);
+    color: var(--text);
+    font-family: var(--fb);
+    font-size: var(--font-sm);
+    outline: none;
+    transition: border-color 0.15s;
+}
+
+.find-input::placeholder { color: var(--dim); }
+.find-input:focus { border-color: var(--cyan); }
+
+.find-btn {
+    padding: 8px 16px;
+    border-radius: 8px;
+    border: none;
+    background: var(--cyan);
+    color: #000;
+    font-family: var(--fb);
+    font-size: var(--font-sm);
+    font-weight: 700;
+    cursor: pointer;
+    white-space: nowrap;
+    transition: box-shadow 0.2s;
+}
+
+.find-btn:hover { box-shadow: 0 0 16px rgba(0, 229, 255, 0.4); }
