+++ client/src/pages/HomePage/SubviewsHomePage/FriendsView/TabViews/FriendsFindTabView/useFriendsFindTabView.ts
@@ -1,0 +1,9 @@
+import { useState } from 'react'
+
+export function useFriendsFindTabView() {
+    const [searchQuery, setSearchQuery] = useState('')
+
+    const handleSendRequest = (username: string) => { console.log('request', username) }
+
+    return { searchQuery, setSearchQuery, handleSendRequest }
+}
