+++ client/src/pages/HomePage/SubviewsHomePage/FriendsView/TabViews/FriendsListTabView/FriendsListTabView.css
@@ -1,0 +1,28 @@
+.friends-grid {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 10px;
+}
+
+.friend-item {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    padding: 8px 10px;
+    border-radius: 10px;
+    background: rgba(255, 255, 255, 0.04);
+    border: 1px solid var(--pb);
+}
+
+.friend-remove {
+    background: transparent;
+    border: none;
+    color: var(--dim);
+    font-family: var(--fb);
+    font-size: 11px;
+    cursor: pointer;
+    transition: color 0.15s;
+    white-space: nowrap;
+}
+
+.friend-remove:hover { color: #f87171; }
